+++ src/drivers/drv_oracledb.ts
@@ -9,10 +9,10 @@
 * @preserve
 */
 'use strict';
-import {MQDriver}  from './index';
-import {MQConst}   from '../mq_const.js';
-import {MQTrace}   from '../mq_trace.js';
-import OracleDB    from 'oracledb';
+import {MQDriver}   from './index';
+import {MQConst}    from '../mq_const.js';
+import {MQTrace}    from '../mq_trace.js';
+import OracleDB     from 'oracledb';
 
 let _initialized = false;
 
@@ -165,13 +165,22 @@ export namespace DrvOracleDB
         // 1. A transaction automatically starts when a DML(INSERT/UPDATE/DELETE/MERGE) statement is executed.
         // 2. Use SET TRANSACTION – this will be handled later as an argument to beginTransaction().
         beginTransaction(): Promise<any> { this.isTR = true; return Promise.resolve(); }
-        query(sql: string): Promise<any> {
-            MQTrace.log(`[C:${this.coid}] [${this.owner.getType()}}]: query:`, sql);
-            if( this.isTR != true ) {
-                // Use autoCommit when not in a transaction-function.
-                return this.conn.execute(sql, {}, { autoCommit: true });
-            }
-            return this.conn.execute(sql);
+        query(sql: string): Promise<MQDriver.ResultSet> {
+            MQTrace.log(`[C:${this.coid}] [${this.owner.getType()}]: query:`, sql);
+
+            const options: OracleDB.ExecuteOptions = {
+                outFormat: OracleDB.OUT_FORMAT_OBJECT,
+                autoCommit: this.isTR !== true,
+            };
+
+            return this.conn.execute(sql, {}, options).then(function(result) {
+                return {
+                    insertId: result.lastRowid,
+                    affected: result.rowsAffected || 0,
+                    rows: result.rows || [], // oracledb.OUT_FORMAT_OBJECT 설정 필수
+                    meta: result,
+                };
+            });
         }
 
         commit(): Promise<any> {